requirements.txt
bleak>=0.20.0
pycryptodome>=3.17.0
flask-socketio>=5.3.0
flask>=2.3.0
lzstring>=1.0.4
